+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GooglePublisher.scala
@@ -15,6 +15,7 @@ trait GooglePublisher[F[_]] {
     * Watch out message size quota and limitations https://cloud.google.com/pubsub/quotas
     */
   def publishString: Pipe[F, String, Unit]
+
 }
 
 object GooglePublisher {